// libs/platform/list/access-data/src/lib/enums/statusses/offer-request-status.enum.ts
export enum OfferRequestStatusEnum {
    Approved = 'approved',
    Pending = 'pending',
    Rejected = 'rejected'
}

export enum OfferRequestAvailabilityEnum {
    Approved = 'approved',
    Rejected = 'rejected'
}

export enum OfferRequestStatusTranslate {
    Approved = 'offers_requests_page.statuses.approved',
    Pending = 'offers_requests_page.statuses.pending',
    Rejected = 'offers_requests_page.statuses.rejected'
}