// libs/platform/list/access-data/src/lib/enums/statusses/scaleo-status.enum.ts
export enum ScaleoStatusEnum {
    Statuses = 'statuses',
    Conversion = 'conversion_statuses',
    Postback = 'postback_statuses',
    Adjustments = 'adjustments_statuses',
    CustomParams = 'custom_params_statuses',
    PaymentsMethods = 'payments_methods_statuses',
    Payment = 'payment_statuses',
    Offers = 'offers_statuses',
    Goals = 'goals_statuses',
    OfferLandingPage = 'offer_urls_statuses',
    Creatives = 'creatives_statuses',
    Invoices = 'invoices_statuses',
    Delivery = 'delivery_status',
    OfferRequest = 'offer_request_statuses',
    OfferAvailability = 'offers_availability',
    Announcement = 'announcement_statuses'
}

export enum ScaleoStatusColorEnum {
    Green = '#27AE60',
    Red = '#f65a63',
    Orange = '#F3B75A',
    Gray = '#b9c6d2'
}